Fix docker removing multiple defs of redismodule.h
