+++ src/redis-roaring.c
@@ -1,6 +1,5 @@
 #include "redismodule.h"
 #include "data-structure.h"
-#include "type.h"
 #include "sds.h"
 
 #define malloc(size) RedisModule_Alloc(size)
@@ -11,6 +10,66 @@
 #include "roaring.c"
 
 static RedisModuleType* BitmapType;
+
+#define BITMAP_ENCODING_VERSION 1
+
+void BitmapRdbSave(RedisModuleIO* rdb, void* value);
+void* BitmapRdbLoad(RedisModuleIO* rdb, int encver);
+void BitmapAofRewrite(RedisModuleIO* aof, RedisModuleString* key, void* value);
+size_t BitmapMemUsage(const void* value);
+void BitmapFree(void* value);
+
+/* === Bitmap type methods === */
+void BitmapRdbSave(RedisModuleIO* rdb, void* value) {
+  Bitmap* bitmap = value;
+  size_t serialized_max_size = roaring_bitmap_size_in_bytes(bitmap);
+  char* serialized_bitmap = RedisModule_Alloc(serialized_max_size);
+  size_t serialized_size = roaring_bitmap_serialize(bitmap, serialized_bitmap);
+  RedisModule_SaveStringBuffer(rdb, serialized_bitmap, serialized_size);
+  RedisModule_Free(serialized_bitmap);
+}
+
+void* BitmapRdbLoad(RedisModuleIO* rdb, int encver) {
+  if (encver != BITMAP_ENCODING_VERSION) {
+    RedisModule_LogIOError(rdb, "warning", "Can't load data with version %d", encver);
+    return NULL;
+  }
+  size_t size;
+  char* serialized_bitmap = RedisModule_LoadStringBuffer(rdb, &size);
+  Bitmap* bitmap = roaring_bitmap_deserialize(serialized_bitmap);
+  RedisModule_Free(serialized_bitmap);
+  return bitmap;
+}
+
+typedef struct Bitmap_aof_rewrite_callback_params_s {
+  RedisModuleIO* aof;
+  RedisModuleString* key;
+} Bitmap_aof_rewrite_callback_params;
+
+static bool BitmapAofRewriteCallback(uint32_t offset, void* param) {
+  Bitmap_aof_rewrite_callback_params* params = param;
+  RedisModule_EmitAOF(params->aof, "R.SETBIT", "sll", params->key, offset, 1);
+  return true;
+}
+
+void BitmapAofRewrite(RedisModuleIO* aof, RedisModuleString* key, void* value) {
+  Bitmap* bitmap = value;
+  Bitmap_aof_rewrite_callback_params params = {
+      .aof = aof,
+      .key = key
+  };
+  roaring_iterate(bitmap, BitmapAofRewriteCallback, &params);
+}
+
+size_t BitmapMemUsage(const void* value) {
+  const Bitmap* bitmap = value;
+  return roaring_bitmap_size_in_bytes(bitmap);
+}
+
+void BitmapFree(void* value) {
+  bitmap_free(value);
+}
+
 
 /* === Bitmap type commands === */
